+++ src/org/graphstream/ui/swingViewer/Selection.java
@@ -31,5 +31,5 @@
 package org.graphstream.ui.swingViewer;
 
 public class Selection {
-	public float x1, y1, x2, y2;
+	public double x1, y1, x2, y2;
 }
